Revert "back to f64"
This reverts commit 7580b8c33305c1bc69db0bbdfe67ac66458237b7.

+++ rust/src/backend.rs
@@ -59,7 +59,7 @@ struct HistoryInfo {
     allInAmount: Amount,
     whoseTurn: String,
     stacks: StacksJSON,
-    winnings: f64,
+    winnings: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
